Update mob_class.py
Updated mob parent class' "on_death" to work with how I have written code in core_game_loop (not published just yet as of 1:33 PM 12/12/16) to handle generation, drawing, and clearing of enemies.  Added "index_update" method. Three arguments added to parent class call, "sprite_kill_list" which will need to be called with (insert sprite type: "minion", "striker", or "turret")_kill_list, "type_killed_num" which is a tracker of how many the player has killed for the level end, and "indexnum" which is the index location of the mob in the list that is used in core_game_loop to keep track of the enemies. The player mob will not supply any of these three arguments, and the asteroid will only need type_killed_num. The three other types of mobs will all need to supply all three of the new arguments.